:truck: rename: Fixture (NAILCASE-96)
- Factory 는 모으는 곳만

resolving: [NAILCASE-96](https://nailcase.atlassian.net/browse/NAILCASE-96)
ref: [NAILCASE-94](https://nailcase.atlassian.net/browse/NAILCASE-94)
