+++ src/test/java/com/nailcase/testUtils/fixture/ShopFixture.java
@@ -1,4 +1,4 @@
-package com.nailcase.testUtils.fixtureFactory;
+package com.nailcase.testUtils.fixture;
 
 import static org.jeasy.random.FieldPredicates.*;
